split stuff out of main
